+++ src/GoogleTestSuiteInfo.ts
@@ -64,8 +64,6 @@ export class GoogleTestSuiteInfo extends AbstractTestSuiteInfo {
             '',
             undefined,
             undefined,
-            this.execPath,
-            this.execOptions,
           );
           super.addChild(test);
           this._shared.sendTestEventEmitter.fire([
@@ -107,8 +105,8 @@ export class GoogleTestSuiteInfo extends AbstractTestSuiteInfo {
 
             for (let j = 0; j < xml.testsuites.testsuite[i].testcase.length; j++) {
               const test = xml.testsuites.testsuite[i].testcase[j];
-              const testName = test.$.name;
-              const testNameFull = suiteName + '.' + testName;
+              const testName = test.$.name.startsWith('DISABLED_') ? test.$.name.substr(9) : test.$.name;
+              const testNameFull = suiteName + '.' + test.$.name;
               let valueParam: string | undefined = undefined;
               if (test.$.hasOwnProperty('value_param')) valueParam = test.$.value_param;
 
@@ -123,8 +121,6 @@ export class GoogleTestSuiteInfo extends AbstractTestSuiteInfo {
                   valueParam,
                   this._findFilePath(test.$.file),
                   test.$.line - 1,
-                  this.execPath,
-                  this.execOptions,
                 ),
               );
             }
@@ -174,9 +170,9 @@ export class GoogleTestSuiteInfo extends AbstractTestSuiteInfo {
             while (testMatch) {
               lineNum++;
 
-              const testName = testMatch[1];
+              const testName = testMatch[1].startsWith('DISABLED_') ? testMatch[1].substr(9) : testMatch[1];
               const valueParam: string | undefined = testMatch[3];
-              const testNameFull = testGroupNameWithDot + testName;
+              const testNameFull = testGroupNameWithDot + testMatch[1];
 
               const old = this.findTestInfoInArray(oldGroupChildren, v => v.testNameFull === testNameFull);
 
@@ -189,8 +185,6 @@ export class GoogleTestSuiteInfo extends AbstractTestSuiteInfo {
                   valueParam,
                   undefined,
                   undefined,
-                  this.execPath,
-                  this.execOptions,
                 ),
               );
 
@@ -266,7 +260,6 @@ export class GoogleTestSuiteInfo extends AbstractTestSuiteInfo {
 
                 data.group = group;
                 this._shared.testStatesEmitter.fire({ type: 'suite', suite: group, state: 'running' });
-                group.sendSkippedChildrenEvents();
               }
             } else {
               this._shared.log.error('should have found group', this, groupName);